+++ src/main/java/com/socialmediaweb/socialmediaweb/dto/SearchResultDTO.java
@@ -3,7 +3,7 @@ package com.socialmediaweb.socialmediaweb.dto;
 import java.util.List;
 
 public class SearchResultDTO {
-    private int userId;
+    private Long userId;
     private String username;
     private String firstName;
     private String lastName;
@@ -20,8 +20,8 @@ public class SearchResultDTO {
     private boolean isFollowing;
     
     // Getters and Setters
-    public int getUserId() { return userId; }
-    public void setUserId(int userId) { this.userId = userId; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
     
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
